do not load
This version is a complete re-implimentation of the Recurrence rule engine.  This change is applied to code that was at least 2 versions old.

The Recurrence rule engine has been reimplimented and appears to be working, but none of the interfacing code has been change yet (so you can't actually use it :)

After the code in ICRecurrenceRule and ICViewableCalanderEntry are fixed I can merge a usable version.

Co-authored-by: Jason <jason@johnson-net.ch>